Add files via upload
This includes the presentation and visualizations, interpreting the models using histograms, scatter plots etc. to uncover the reason for the higher number of diabetic patients amongst the PIMA Indian women. 

Predicting whether a patient would be diagnosed with diabetes or not based on parameters available in the dataset

Breaking down the relationships between different parameters of the chosen dataset and find a correlations to better understand the end result.